requirements.txt
asgiref==3.8.1
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.0.6
django-environ==0.12.0
django-jalali==7.4.0
django-storages==1.14.6
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
jmespath==1.0.1
pillow==11.2.1
python-dateutil==2.9.0.post0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
smsir-python==1.0.8
django-ckeditor==6.5.1